[branch dev-btm]: CHECKPOINT FOR SAFETY - phase 1 of callable executor (client service) smart proxy work. Includes fixes to issues identified by the tests; in particular the com.bigdata.service and com.bigdata.jini tests
git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/dev-btm@3919 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata-jini/src/java/com/bigdata/metadata/Constants.java
@@ -98,7 +98,7 @@ public interface Constants {
          (IBigdataClient.Options.DEFAULT_CLIENT_MAX_PARALLEL_TASKS_PER_REQUEST);
 
     long LOWER_BOUND_TASK_TIMEOUT = 0;
-    long UPPER_BOUND_TASK_TIMEOUT = Integer.MAX_VALUE;
+    long UPPER_BOUND_TASK_TIMEOUT = Long.MAX_VALUE;
     long DEFAULT_TASK_TIMEOUT =
          Long.parseLong
          (IBigdataClient.Options.DEFAULT_CLIENT_TASK_TIMEOUT);
